moved schema closed to the lambda function definition and documentation. Other small fixes
